[GTK][WPE] missing URI for injected user scripts and style sheets https://bugs.webkit.org/show_bug.cgi?id=272152
Reviewed by Adrian Perez de Castro.

User scripts/style sheets without the URI are not visible in WebInspector.

With this change the URI of injected user scripts/style sheets are generated as:
user-script:incremented_identifier
user-style:incremented_identifier

* LayoutTests/userscripts/window-onerror-for-isolated-world-1-expected.txt:
* LayoutTests/userscripts/window-onerror-for-isolated-world-2-expected.txt:
* LayoutTests/userscripts/window-onerror-for-isolated-world-3-expected.txt:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/page/UserScript.cpp: Copied from Source/WebKit/UIProcess/API/APIUserScript.cpp.
(WebCore::generateUniqueURL):
(WebCore::UserScript::UserScript):
* Source/WebCore/page/UserScript.h:
(WebCore::UserScript::UserScript): Deleted.
* Source/WebCore/page/UserStyleSheet.cpp: Copied from Source/WebKit/UIProcess/API/APIUserScript.cpp.
(WebCore::generateUniqueURL):
(WebCore::UserStyleSheet::UserStyleSheet):
* Source/WebCore/page/UserStyleSheet.h:
* Source/WebCore/page/WorkerNavigator.cpp:
* Source/WebKit/UIProcess/API/APIUserScript.cpp:
(API::UserScript::generateUniqueURL): Deleted.
* Source/WebKit/UIProcess/API/APIUserScript.h:
* Source/WebKit/UIProcess/API/APIUserStyleSheet.cpp:
(API::UserStyleSheet::generateUniqueURL): Deleted.
* Source/WebKit/UIProcess/API/APIUserStyleSheet.h:
* Source/WebKit/UIProcess/API/C/WKUserScriptRef.cpp:
(WKUserScriptCreateWithSource):

Canonical link: https://commits.webkit.org/277470@main
